ziad

+++ app/Http/Controllers/ProjectController.php
@@ -191,16 +191,22 @@ class ProjectController extends Controller
     {
         try {
             $project= Project::find($id);
+       $benefactoryPoject = BeneficiariesProject::where('project_id','=',$id)->get();
             if (!$project){
                 toastr()->error(__('يوجد خطاء ما'));
-                return redirect()->route('user.index');
+                return redirect()->route('projects.index');
+            }
+            if (count($benefactoryPoject)>0){
+                toastr()->error(__('هذا المشروع يحتوي على مستفيدين لا يمكن حذفه'));
+                return redirect()->route('projects.index');
             }
             $project ->delete();
 
-            toastr()->success(__('تم تحديث البيانات بنجاح'));
+            toastr()->success(__('تم حذف المشروع بنجاح'));
             return back();
 
         }catch (\Exception $ex){
+            return $ex;
             toastr()->error(__('يوجد خطاء ما'));
             return back();
         }
